[유용민] layouts 폴더 생성 및 분리
