+++ app/[id]/layout.tsx
@@ -1,18 +1,9 @@
-import cn from "classnames/bind";
-import styles from "./layout.module.scss";
+import DetailLayout from "@/layouts/detailLayout/DetailLayout";
 
-const cx = cn.bind(styles);
-
-export default function DetailLayout({
+export default function Layout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  return (
-    <div className={cx("pageWrapper")}>
-      <div className={cx("contentContainer")}>
-        <div className={cx("content")}>{children}</div>
-      </div>
-    </div>
-  );
+  return <DetailLayout>{children}</DetailLayout>;
 }
