+++ assets/icons/Chat.tsx
@@ -8,8 +8,8 @@ export default function Chat() {
       xmlns="http://www.w3.org/2000/svg"
     >
       <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
+        fillRule="evenodd"
+        clipRule="evenodd"
         d="M8.99984 0.5C4.39799 0.5 0.666504 3.43515 0.666504 7.05484C0.666504 9.30544 2.11095 11.2935 4.30539 12.4656L3.37947 15.9072C3.29613 16.2072 3.63873 16.4511 3.89799 16.2729L7.95354 13.5534C8.29613 13.5909 8.63873 13.6097 8.99984 13.6097C13.6017 13.6097 17.3332 10.6745 17.3332 7.05484C17.3332 3.43515 13.6017 0.5 8.99984 0.5Z"
         fill="#121416"
       />
