+++ layouts/detailLayout/DetailLayout.module.scss
