+++ layouts/mainLayout/MainLayout.tsx
@@ -1,0 +1,7 @@
+export default function MainLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <>{children}</>;
+}
